requirements.txt
streamlit
python-dotenv
langchain
langchain_openai